+++ utoipa/src/lib.rs
@@ -81,6 +81,10 @@
 //!   By default these types are parsed as `string`. `OffsetDateTime` and `PrimitiveDateTime` will use `date-time` format. `Date` will use
 //!   `date` format and `Duration` will not have any format. To override default `string` representation users have to use `value_type` attribute
 //!   to override the type. See [docs](https://docs.rs/utoipa/latest/utoipa/derive.ToSchema.html) for more details.
+//! * **`jiff_0_2`** Add support for [jiff 0.2](https://crates.io/crates/jiff) `Zoned`, and `civil::Date` types.
+//!   By default these types are parsed as `string`. `Zoned` will use `date-time` format. `civil::Date` will use
+//!   `date` format. To override default `string` representation users have to use `value_type` attribute
+//!   to override the type. See [docs](https://docs.rs/utoipa/latest/utoipa/derive.ToSchema.html) for more details.
 //! * **`decimal`** Add support for [rust_decimal](https://crates.io/crates/rust_decimal) `Decimal` type. **By default**
 //!   it is interpreted as `String`. If you wish to change the format you need to override the type.
 //!   See the `value_type` in [`ToSchema` derive docs][to_schema_derive].
@@ -608,9 +612,9 @@ where
 
 #[cfg(feature = "macros")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "macros")))]
-impl<K: ToSchema, T: ToSchema> ToSchema for std::collections::HashMap<K, T>
+impl<K: ToSchema, T: ToSchema, S> ToSchema for std::collections::HashMap<K, T, S>
 where
-    std::collections::HashMap<K, T>: PartialSchema,
+    std::collections::HashMap<K, T, S>: PartialSchema,
 {
     fn schemas(
         schemas: &mut Vec<(
@@ -642,9 +646,9 @@ where
 
 #[cfg(feature = "macros")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "macros")))]
-impl<K: ToSchema> ToSchema for std::collections::HashSet<K>
+impl<K: ToSchema, S> ToSchema for std::collections::HashSet<K, S>
 where
-    std::collections::HashSet<K>: PartialSchema,
+    std::collections::HashSet<K, S>: PartialSchema,
 {
     fn schemas(
         schemas: &mut Vec<(
@@ -1437,7 +1441,7 @@ pub mod __dev {
         }
     }
 
-    impl<K: ComposeSchema, T: ComposeSchema> ComposeSchema for std::collections::HashMap<K, T> {
+    impl<K: ComposeSchema, T: ComposeSchema, S> ComposeSchema for std::collections::HashMap<K, T, S> {
         fn compose(
             schemas: Vec<utoipa::openapi::RefOr<utoipa::openapi::schema::Schema>>,
         ) -> utoipa::openapi::RefOr<utoipa::openapi::schema::Schema> {
@@ -1459,7 +1463,7 @@ pub mod __dev {
         }
     }
 
-    impl<K: ComposeSchema> ComposeSchema for std::collections::HashSet<K> {
+    impl<K: ComposeSchema, S> ComposeSchema for std::collections::HashSet<K, S> {
         fn compose(
             schemas: Vec<utoipa::openapi::RefOr<utoipa::openapi::schema::Schema>>,
         ) -> utoipa::openapi::RefOr<utoipa::openapi::schema::Schema> {