+++ utoipa-gen/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog - utoipa-gen
 
+## 5.4.0 - Jun 16 2025
+
+### Added
+
+* Add feature support extensions in `utoipa::path` macro (https://github.com/juhaku/utoipa/pull/1292)
+* Add support for jiff v0.2 (https://github.com/juhaku/utoipa/pull/1332)
+
 ## 5.3.1 - Jan 6 2025
 
 ### Fixed